Test File Consolidation and Naming Standardization (#5865)
